Iterate over cell types
In order to assist with transfering boundary data between blocks,
an iterator has been created that can iterate over different cell types.
In addition, a method to collect all the values of a particular type
and return an owned vector has been writte, and this will serve as the
means to transfer the boundary data between blocks.